fixed sigmoid function

+++ neural_network.java
@@ -2,8 +2,8 @@ public class neural_network{
     static double[] x = {4.0,5.0};
     double[] weighths={3.0,3.0};
     double bias = 1.0;
-    public int sigmoid (double z){
-        return (int)(1/(1+Math.exp(-z)));
+    public double  sigmoid (double z){
+        return (1/(1+Math.exp(-z)));
     }
     public static void main(String[] args){
         neural_network nn = new neural_network();
